+++ front-end/tests/tests.rs
@@ -1,6 +1,7 @@
 mod utils;
 
 mod config;
+mod cursor;
 mod encoding;
 mod file_discovery;
 mod help;